wallet: improve naming, remove redundant WalletStock
